requirements.txt
httpx>=0.28.1
mcp[cli]>=1.2.1
pillow>=11.1.0
azure-storage-blob>=12.26.0
azure-identity>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
selenium>=4.0.0
starlette>=0.27.0
sseclient-py>=1.8.0